test: quaternion unit tests
There's an interesting issue in finding the mean of quaternions
when close to pi. The algorithm converges, but it doesn't converge
to the right value, i.e. it finds a local minimum.

However, if an initial value is provided that is close enough to
the resulting mean, the correct mean is found.

This suggests that the gradient descent algorithm needs to get
fancier, or a different algorithm needs to be chosen.