feat: more vicon and imu data
Oddly enough, the fixes I made to `to_vectors` and `find_q_mean`
seem to cause issues when filtering certain datasets. For example,
when there's no vicon data, the predicted yaw goes a wild.

This is likely indicative of an issue with the filtering algorithm
though I don't know yet what that is.