feat: trainer
Trains coefficients for accelerometer and gyro data with linear
regression.

The rotation matrices can be used to create training data for both
cases. There are some inconsistencies in the gyro data for the
first dataset. When there should be a spike in angular velocity
near the beginning, the gyro data is flat.