fix: Quaternions.to_vectors() clipping
At long last, discovered the bug that I think was dogging everyone when
it came to making the filter stable.

When calling to_vectors() on an array of quaternions, the output needs
adjusted to be between -pi and pi.

A stray vector with one value close to -2*pi was sending everything into
disarray before, but with that solved, I can pick reasonable values for
Q and R without tuning to the .001 decimal place.